Clean up user context and modals
